Move lokalise plugin extension to its own package
